fixed homework19
