+++ src/main/java/com/practicesoftwaretesting/cart/model/CreatedCartResponse.java
@@ -1,4 +1,4 @@
-package com.practicesoftwaretesting.Cart.Model;
+package com.practicesoftwaretesting.cart.model;
 
 import lombok.Data;
