+++ src/main/java/com/practicesoftwaretesting/user/model/NewUserRegisterRequest.java
@@ -1,14 +1,13 @@
-package com.practicesoftwaretesting.User.Model;
+package com.practicesoftwaretesting.user.model;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Builder;
 import lombok.Data;
 
 @Data
-public class NewUserRegisteredResponse {
+@Builder
+public class NewUserRegisterRequest {
 
-    @JsonProperty("first_name")
     private String firstName;
-    @JsonProperty("last_name")
     private String lastName;
     private String address;
     private String city;
@@ -17,8 +16,6 @@ public class NewUserRegisteredResponse {
     private String postcode;
     private String phone;
     private String dob;
+    private String password;
     private String email;
-    private String id;
-    @JsonProperty("created_at")
-    private String createdAt;
 }
