+++ src/main/java/com/practicesoftwaretesting/user/model/UserLoginResponse.java
@@ -1,4 +1,4 @@
-package com.practicesoftwaretesting.User.Model;
+package com.practicesoftwaretesting.user.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.Data;
