+++ src/test/java/com/practicesoftwaretesting/api/BaseTest.java
@@ -1,4 +1,4 @@
-package com.practicesoftwaretesting.APItests;
+package com.practicesoftwaretesting.api;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
